+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [0.2.0]
+- Added Hierarchical clustering
+- Moved euclidDistance to common file
+
 ## [0.1.2]
 - Updated License